requirements.txt
RPi.GPIO
transformers
torch
twilio
mysql-connector-python